sales_sheets_to_bq/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.10.0
pandas==1.5.0
numpy==1.23.2
google-cloud-storage==2.5.0
openpyxl
requests==2.28.1
pyarrow==14.0.1